requirements.txt
asgiref==3.6.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==3.2.18
django-filter==23.2
djangorestframework==3.14.0
importlib-metadata==6.6.0
ipython==7.34.0
jedi==0.18.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
Pygments==2.15.1
pytz==2023.3
sqlparse==0.4.4
traitlets==5.9.0
typing-extensions==4.5.0
wcwidth==0.2.6
zipp==3.15.0
